feat(ci): speed up builds by accepting browser as argument

+++ Justfile
@@ -26,7 +26,7 @@ lint: build-prod
     web-ext lint --source-dir=.build_firefox
 
 # Build extension
-_build:
+_build BROWSER="all":
     #!/usr/bin/env nu
     # workaround for vite's build process that triggers the wrong code path for sanctuary modules
     # This issue has been fixed in currently unrelease sanctuary versions
@@ -49,31 +49,35 @@ _build:
     let build_dir = $env.BUILD_DIR
     rm -rpf $build_dir
     mkdir $build_dir
-    ## frontend build
-    yarn build
-    ^cp -r ./.output/public/* $build_dir # solid-start always builds everything in the dist directory
-    rm -rvpf $"($build_dir)/assets"
-    ^find $build_dir -name '*.gz' -exec rm -v {} +
-    ^find $build_dir -name '*.br' -exec rm -v {} +
-    ^find $build_dir -name '*.wasm' -exec rm -v {} +
-    ^find $build_dir -name '*.json' -exec rm -v {} +
-    mv $"($build_dir)/index" $"($build_dir)/index.html"
-    # INFO: workaround for https://github.com/solidjs/solid-start/issues/1263
-    htmlq -f $"($build_dir)/index.html" -o $"($build_dir)/manifest.js" --text 'body > script:first-of-type'
-    htmlq -f $"($build_dir)/index.html" -r 'body > script:first-of-type' | sed -e 's#</div>#</div><script src="/manifest.js"></script>#' | save -f $"($build_dir)/index.html.new"
-    mv -f $"($build_dir)/index.html.new" $"($build_dir)/index.html"
-    ## backend build
-    yarn run rollup -c
-    # ls *.js | par-each {|it| yarn run rollup -i $it.name --file $"($build_dir)/($it.name | path basename)" --format iife --inlineDynamicImports -p @rollup/plugin-commonjs -p rollup-plugin-polyfill-node -p @rollup/plugin-node-resolve}
-    # rmdir $"($build_dir)/.solid"
-    mv .build_background.js .build/background.js
+    [rollup frontend] | par-each {|type|
+      if $type == "rollup" {
+        ## backend build
+        yarn run rollup -c
+        mv .build_background.js .build/background.js
+        # ls *.js | par-each {|it| yarn run rollup -i $it.name --file $"($build_dir)/($it.name | path basename)" --format iife --inlineDynamicImports -p @rollup/plugin-commonjs -p rollup-plugin-polyfill-node -p @rollup/plugin-node-resolve}
+      } else {
+        ## frontend build
+        yarn build
+        ^cp -r ./.output/public/* $build_dir # solid-start always builds everything in the dist directory
+        rm -rvpf $"($build_dir)/assets"
+        ^find $build_dir -name '*.gz' -exec rm -v {} +
+        ^find $build_dir -name '*.br' -exec rm -v {} +
+        ^find $build_dir -name '*.wasm' -exec rm -v {} +
+        ^find $build_dir -name '*.json' -exec rm -v {} +
+        # INFO: workaround for https://github.com/solidjs/solid-start/issues/1263
+        htmlq -f $"($build_dir)/index.html" -o $"($build_dir)/manifest.js" --text 'body > script:first-of-type'
+        htmlq -f $"($build_dir)/index.html" -r 'body > script:first-of-type' | sed -e 's#</div>#</div><script src="/manifest.js"></script>#' | save -f $"($build_dir)/index.html.new"
+        mv -f $"($build_dir)/index.html.new" $"($build_dir)/index.html"
+      }
+    }
+
     ## prepare additional files
     let dist_dir = $env.DIST_DIR
     rm -rpf $dist_dir
     mkdir $dist_dir
     cp LICENSE $build_dir
     ## package plugin
-    [firefox chrome source] | par-each {|browser|
+    if "{{ BROWSER }}" == "all" {[firefox chrome source]} else {"{{ BROWSER }}" | split words} | par-each {|browser|
       let build_dir_browser = $"($env.BUILD_DIR)_($browser)"
       rm -rpf $build_dir_browser
       cp -r $build_dir $build_dir_browser
@@ -93,8 +97,10 @@ _build:
         mv $"($build_dir_browser).crx" $"($dist_dir)/($manifest_shared.name)-($manifest_shared.version)_($browser).crx"
         cd $build_dir_browser
         ^zip -q -r -0 $"../($dist_dir)/($manifest_shared.name)-($manifest_shared.version)_($browser).zip" *
-      } else {
+      } else if $browser == "source" {
         git archive --format=zip HEAD -o $"($dist_dir)/($manifest_shared.name)-($manifest_shared.version)_($browser).zip"
+      } else {
+        print -e $"Unknown browser, only source, firefox, chrome and all are supported: ($browser)"
       }
       print $"($browser) package ready: ($dist_dir)/($manifest_shared.name)-($manifest_shared.version)_($browser).zip"
       $"($dist_dir)/($manifest_shared.name)-($manifest_shared.version)_($browser).zip"
@@ -102,26 +108,26 @@ _build:
     print "done."
 
 # Build extension for development
-build-dev:
-    NODE_ENV=development just _build
+build-dev BROWSER="all":
+    NODE_ENV=development just _build {{ BROWSER }}
 
 # Build extension for production
-build-prod:
-    NODE_ENV=production just _build
+build-prod BROWSER="all":
+    NODE_ENV=production just _build {{ BROWSER }}
 
-_build-notify:
+_build-notify BROWSER="all":
     #!/usr/bin/env nu
     let start = (date now)
-    just _build
+    NODE_ENV=development just _build {{ BROWSER }}
     notify-send -a identinet-plugin $"(date now | format date "%H:%M") - built, duration: ((date now) - $start)"
 
 # Watch changes and rebuild appliaction
-build-watch:
+build-watch BROWSER="all":
     # FIXME: this isn't optimal - not all files are being watched,
     # ./background.js and ./public are missing. Furthermore, it would be great
     # to perform the build when the task is started
     # watch src {|| let start = (date now); just build; notify-send -a identinet-plugin $"(date now | format date "%H:%M:%S") - build complete - it took ((date now) - $start)"}
-    watchexec -r -w src -w ./Justfile -w ./background.js -w ./public -w ./package.json -w ./vite.config.js -w ./rollup.config.js -w ./uno.config.ts -w ./manifest just _build-notify
+    watchexec -r -w src -w ./Justfile -w ./background.js -w ./public -w ./package.json -w ./vite.config.js -w ./rollup.config.js -w ./uno.config.ts -w ./manifest just _build-notify {{ BROWSER }}
 
 # Run local test websites
 run-websites:
